Refactor deploy method to deploy private repos using deploy keys(WIP)
