+++ app/models/server.rb
@@ -1,14 +1,12 @@
 require 'net/ssh'
 
 class Server < ApplicationRecord
+  include SSHKeyGenerator
   has_many :apps
   has_many :plugin_instances
   has_many :ssh_keys
   belongs_to :user
 
-  attr_accessor :generate_keys
-
-  before_save :generate_and_save_keys
   before_save :parse_private_key
 
   def execute(cmd)
@@ -93,13 +91,6 @@ class Server < ApplicationRecord
     self.name
   end
 
-  def generate_and_save_keys
-    if generate_keys == "1"
-      key = SSHKey.generate
-      self.private_key = key.private_key
-      self.public_key = key.ssh_public_key
-    end
-  end
 
   def parse_private_key
     self.private_key = self.private_key.gsub("\r\n          ", "\n") #TODO find better way
