+++ app/views/apps/show.html.haml
@@ -42,8 +42,8 @@
         %dl.dl-horizontal
           %dt Name:
           %dd= @app.name
-          %dt Url:
-          %dd= @app.url
+          %dt Git Url:
+          %dd= @app.git_url
           %dt Server:
           %dd= @app.server
