+++ db/migrate/20160804144126_add_keys_to_apps.rb
@@ -1,0 +1,6 @@
+class AddKeysToApps < ActiveRecord::Migration[5.0]
+  def change
+    add_column :apps, :private_key, :text
+    add_column :apps, :public_key, :text
+  end
+end
